RTS Toolkit
This is piece of software I purchased as a starter kit for people to
play with. The basic pieces are here, just toy around with altering the
mechanics to learn how unreal works. Single unit selection, a basic
camera, and basic AI are here. Perhaps try to create drag box selection,
greater zoom on the camera, and tweaking stats. Have fun, and remember
that this is not the actual project!